# api/openapi3.json
{
    "openapi": "3.0.0",
    "info": {
        "contact": {},
        "description": "API REST para gerenciamento de chamados de suporte",
        "title": "Sistema de Chamados",
        "version": "1.0.0"
    },
    "paths": {
        "/api/assignees": {
            "get": {
                "description": "Retorna todos os atendentes e a contagem de chamados em aberto (status open ou in_progress)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/AssigneeOutput"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Lista todos os responsáveis",
                "tags": [
                    "Assignees"
                ]
            },
            "post": {
                "description": "Registra um novo membro da equipe de suporte",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAssigneeInput"
                            }
                        }
                    },
                    "description": "Dados do Responsável",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Mensagem de sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Cadastra um novo responsável",
                "tags": [
                    "Assignees"
                ]
            }
        },
        "/api/tickets": {
            "get": {
                "description": "Retorna uma lista paginada de chamados de acordo com os filtros especificados",
                "parameters": [
                    {
                        "description": "Número da página (padrão: 1)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Itens por página (máx: 100, padrão: 10)",
                        "in": "query",
                        "name": "page_size",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Campo de ordenação (created_at, priority, status, title)",
                        "in": "query",
                        "name": "sort_by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Sentido da ordenação (asc, desc)",
                        "in": "query",
                        "name": "order",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por status (open, in_progress, resolved, closed)",
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por prioridade (low, medium, high)",
                        "in": "query",
                        "name": "priority",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por nome do responsável",
                        "in": "query",
                        "name": "assignee",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Busca textual em título e descrição",
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedTicketsOutput"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Lista chamados com paginação, filtros e ordenação",
                "tags": [
                    "Tickets"
                ]
            },
            "post": {
                "description": "Registra um chamado com atribuição manual ou automática (auto_assign: true)",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTicketInput"
                            }
                        }
                    },
                    "description": "Dados do Chamado",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketOutput"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Cria um novo chamado",
                "tags": [
                    "Tickets"
                ]
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "description": "Retorna os detalhes de um chamado específico pelo seu UUID",
                "parameters": [
                    {
                        "description": "ID do Chamado (UUID)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketOutput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Busca chamado por ID",
                "tags": [
                    "Tickets"
                ]
            },
            "put": {
                "description": "Atualiza apenas o status e/ou o responsável (manual ou automático) de um chamado. O conteúdo do chamado (título e descrição) não pode ser editado.",
                "parameters": [
                    {
                        "description": "ID do Chamado (UUID)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketInput"
                            }
                        }
                    },
                    "description": "Dados para atualização de status/responsável",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketOutput"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Atualiza responsável ou status de um chamado",
                "tags": [
                    "Tickets"
                ]
            }
        },
        "/health": {
            "get": {
                "description": "Retorna o status de integridade do serviço.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "summary": "Checagem de integridade da API",
                "tags": [
                    "Health"
                ]
            }
        }
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "schemas": {
            "AssigneeOutput": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "open_tickets_count": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateAssigneeInput": {
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateTicketInput": {
                "properties": {
                    "assignee": {
                        "type": "string"
                    },
                    "auto_assign": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaginatedTicketsOutput": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/TicketOutput"
                        },
                        "type": "array"
                    },
                    "page_size": {
                        "type": "integer"
                    },
                    "total_items": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "TicketOutput": {
                "properties": {
                    "assignee": {
                        "type": "string"
                    },
                    "assignee_details": {
                        "$ref": "#/components/schemas/AssigneeOutput"
                    },
                    "assignee_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateTicketInput": {
                "properties": {
                    "assignee": {
                        "type": "string"
                    },
                    "auto_assign": {
                        "type": "boolean"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}